Enable Conan on Travis, minor CMake cleanups
* Use BUILD_TESTING CMake variable
* Use Conan on Travis
* Test package compiles and runs tests
