+++ .appveyor.yml
@@ -1,14 +1,53 @@
-image:
-  - Visual Studio 2017
-  
+build: false
+
+environment:
+    CONAN_BUILD_REQUIRES: "cmake_installer/3.14.7@conan/stable, ninja/1.9.0"
+    CONAN_CMAKE_GENERATOR: Ninja
+    CONAN_PRINT_RUN_COMMANDS: 1
+    PYTHON_HOME: "C:\\Python37"
+
+    matrix:
+        - MINGW_CONFIGURATIONS: '6@x86_64@seh@posix'
+        - MINGW_CONFIGURATIONS: '7@x86_64@seh@posix'
+        - MINGW_CONFIGURATIONS: '8@x86_64@seh@posix'
+        - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
+          CONAN_VISUAL_VERSIONS: 15
+          CONAN_BUILD_TYPES: Release
+          CONAN_ARCHS: x86_64
+        - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
+          CONAN_VISUAL_VERSIONS: 15
+          CONAN_BUILD_TYPES: Release
+          CONAN_ARCHS: x86
+        - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
+          CONAN_VISUAL_VERSIONS: 15
+          CONAN_BUILD_TYPES: Debug
+          CONAN_ARCHS: x86_64
+        - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
+          CONAN_VISUAL_VERSIONS: 15
+          CONAN_BUILD_TYPES: Debug
+          CONAN_ARCHS: x86_64
+        - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+          CONAN_VISUAL_VERSIONS: 16
+          CONAN_BUILD_TYPES: Release
+          CONAN_ARCHS: x86_64
+        - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+          CONAN_VISUAL_VERSIONS: 16
+          CONAN_BUILD_TYPES: Release
+          CONAN_ARCHS: x86
+        - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+          CONAN_VISUAL_VERSIONS: 16
+          CONAN_BUILD_TYPES: Debug
+          CONAN_ARCHS: x86_64
+        - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+          CONAN_VISUAL_VERSIONS: 16
+          CONAN_BUILD_TYPES: Debug
+          CONAN_ARCHS: x86_64
+
 install:
-  - git submodule update --init
+  - set PATH=%PYTHON_HOME%;%PYTHON_HOME%/Scripts/;%PATH%
+  - pip.exe install conan --upgrade
+  - pip.exe install conan_package_tools
+  - conan user # It creates the conan data directory
 
-clone_folder: c:\projects\propagate_const
-
-before_build:
-  - cd c:\projects\propagate_const
-  
-build_script:
-  - python scripts\build.py -t
-
+test_script:
+  - python build.py
