+++ test_package/conanfile.py
@@ -1,0 +1,23 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
+
+from conans import ConanFile, CMake, tools
+import os
+
+
+class TestPackageConan(ConanFile):
+    settings = "os", "compiler", "build_type", "arch"
+    generators = "cmake"
+    build_requires = "Catch2/2.10.2@catchorg/stable"
+
+    def build(self):
+        cmake = CMake(self)
+        cmake.configure()
+        cmake.build()
+
+    def test(self):
+    	if tools.cross_building(self.settings):
+    		return
+    	bin_path = os.path.join("bin", "test_propagate_const")
+    	self.run(bin_path, run_environment=True)
